+++ HISTORY.md
@@ -1,5 +1,11 @@
 # Revision History for "analyzefit"
 
+## Revision 0.1.0
+- Fixed the line in the quantile plot so that it is a linear fit of
+  the data and the distribution.
+- Initial implementation of the Spread-location plot added. Need to
+  add line that splits data set in half.
+
 ## Revision 0.0.3
 -Finished implementing Quantile plots. The dashed line going through
  them still needs some work though.